Revert "KPMP-5472_remove-spectrack-connection"
