+++ data_management/DluWatcher
@@ -4,11 +4,7 @@ COPY requirements.txt ./
 
 RUN pip3 install --no-cache-dir -r requirements.txt
 COPY ./lib/ ./lib
-COPY ./services/dlu_filesystem.py ./services/dlu_filesystem.py
-COPY ./services/dlu_package_inventory.py ./services/dlu_package_inventory.py
-COPY ./services/dlu_state.py ./services/dlu_state.py
-COPY ./services/dlu_management.py ./services/dlu_management.py
-COPY ./services/dlu_mongo.py ./services/dlu_mongo.py
+COPY ./services/ ./services
 COPY ./model ./model
 COPY ./watch_files.py ./
