+++ data_management/lib/spectrack_connection.py
@@ -8,6 +8,7 @@ import logging
 
 logger = logging.getLogger("lib-SpectrackConnection")
 logging.basicConfig(level=logging.ERROR)
+
 class SpectrackConnection:
 
     SPECIMEN_URL_SUFFIX = "/specimens"
