+++ data_management/main.py
@@ -1,5 +1,4 @@
-from services.dlu_management import DluManagement
-from services.spectrack_management import SpectrackManagement
+from services.data_management import DataManagement
 from services.redcap import Redcap
 import argparse
 import logging
@@ -10,22 +9,21 @@ logger.setLevel(logging.INFO)
 
 class Main:
     def __init__(self):
-        self.dlu_management = DluManagement()
-        self.spectrack_management = SpectrackManagement()
+        self.data_management = DataManagement()
 
     def import_redcap_data(self):
-        dlu_management = DluManagement()
+        data_management = DataManagement()
         redcap = Redcap()
         redcap.set_redcap_participant_data()
         redcap_participant_data = redcap.get_redcap_participant_data()
         for redcap_participant in redcap_participant_data:
-            dlu_management.insert_redcap_participant(redcap_participant)
+            data_management.insert_redcap_participant(redcap_participant)
 
     def insert_all_spectrack_specimens(self):
-        return self.spectrack_management.insert_all_spectrack_specimens()
+        return self.data_management.insert_all_spectrack_specimens()
 
     def upsert_new_spectrack_specimens(self):
-        return self.spectrack_management.upsert_new_spectrack_specimens()
+        return self.data_management.upsert_new_spectrack_specimens()
 
 
 if __name__ == "__main__":
