+++ data_management/md5_updater.py
@@ -2,7 +2,7 @@ import argparse
 import logging
 from services.dlu_mongo import DLUMongo
 from lib.mongo_connection import MongoConnection
-from services.dlu_management import DluManagement
+from services.data_management import DataManagement
 from dotenv import load_dotenv
 from services.dlu_filesystem import calculate_checksum, DLUFile
 import os
@@ -16,7 +16,7 @@ class Main:
     def __init__(self):
         self.mongo_connection = MongoConnection().get_mongo_connection()
         self.dlu_mongo = DLUMongo(self.mongo_connection)
-        self.dlu_management = DluManagement()
+        self.data_management = DataManagement()
         self.data_lake_directory = os.environ["checker_dlu_data_directory"]
 
     def fill_mongo_missing_md5s(self, report_only: bool = False):
@@ -76,7 +76,7 @@ class Main:
 
     def fill_dmd_missing_md5s(self, report_only: bool = False):
         logger.info("Handling DMD records missing md5checksum")
-        files = self.dlu_management.find_files_missing_md5()
+        files = self.data_management.find_files_missing_md5()
         for file in files:
             if report_only:
                 logger.error(
@@ -84,14 +84,14 @@ class Main:
             else:
                 new_checksum = self.calculate_md5(file_name=file["dlu_fileName"], package_id=file["dlu_package_id"])
                 if new_checksum is not None:
-                    self.dlu_management.update_md5(file["dlu_file_id"], new_checksum, file["dlu_package_id"])
+                    self.data_management.update_md5(file["dlu_file_id"], new_checksum, file["dlu_package_id"])
 
     def fix_dmd_md5s(self, report_only: bool = False, fill_missing_only: bool = False):
         if fill_missing_only:
             self.fill_dmd_missing_md5s(report_only=report_only)
         else:
-            files = self.dlu_management.find_all_files()
             logger.info("Handling DMD records with incorrect md5checksums")
+            files = self.data_management.find_all_files()
             for file in files:
                 checksum = self.calculate_md5(file_name=file["dlu_fileName"], package_id=file["dlu_package_id"])
                 logger.info(checksum)
@@ -105,7 +105,7 @@ class Main:
                                      " incorrect md5")
                 else:
                     if file["dlu_md5checksum"] is None or file["dlu_md5checksum"] != checksum and checksum is not None:
-                        self.dlu_management.update_md5(file["dlu_file_id"], checksum, file["dlu_package_id"])
+                        self.data_management.update_md5(file["dlu_file_id"], checksum, file["dlu_package_id"])
 
     def calculate_md5(self, file_name, package_id):
         full_path = os.path.join(self.data_lake_directory, "package_" + package_id + "/"
