+++ data_management/run.sh
@@ -1,1 +1,1 @@
-docker run --log-driver=none -a stdin -a stdout -a stderr --network dataLake --env-file .env kingstonduo/data-management:1.8 python ./main.py $1 $2 $3 $4
+docker run --log-driver=none -a stdin -a stdout -a stderr --network dataLake --env-file .env kingstonduo/data-management:latest python ./main.py $1 $2 $3 $4
