+++ data_management/watch_files.py
@@ -2,7 +2,7 @@ from lib.mongo_connection import MongoConnection
 from services.dlu_filesystem import DLUFileHandler, DirectoryInfo, DLUFile
 from services.dlu_package_inventory import DLUPackageInventory
 from services.dlu_state import DLUState, PackageState
-from services.dlu_management import DluManagement
+from services.data_management import DataManagement
 from model.dlu_package import DLUPackage
 from services.dlu_mongo import DLUMongo
 
@@ -24,7 +24,7 @@ class DLUWatcher:
             self.db = DLUPackageInventory()
         self.mongo_connection = MongoConnection().get_mongo_connection()
         self.dlu_mongo = DLUMongo(self.mongo_connection)
-        self.dlu_management = DluManagement()
+        self.data_management = DataManagement()
         self.dlu_file_handler = DLUFileHandler()
         self.dluPackage = DLUPackage()
         self.dlu_state = DLUState()
@@ -57,12 +57,12 @@ class DLUWatcher:
             package_id = package['dlu_package_id']
             logger.info("Moving package " + package_id)
 
-            self.dlu_management.update_dlu_package(package_id, { "globus_dlu_status": "processing" })
+            self.data_management.update_dlu_package(package_id, { "globus_dlu_status": "processing" })
             globus_data_directory = '/globus/' + package_id
             if not os.path.isdir(globus_data_directory):
                 error_msg = "Error: package " + package_id + " not found in directory " + globus_data_directory + "."
                 logger.info(error_msg + " Skipping.")
-                self.dlu_management.update_dlu_package(package_id, { "globus_dlu_status": error_msg })
+                self.data_management.update_dlu_package(package_id, { "globus_dlu_status": error_msg })
                 continue
 
             directory_info = DirectoryInfo(globus_data_directory)
@@ -70,7 +70,7 @@ class DLUWatcher:
             if directory_info.file_count == 0 and directory_info.subdir_count == 0:
                 error_msg = "Error: package " + package_id + " has no files or top level subdirectory"
                 logger.info(error_msg + " Skipping.")
-                self.dlu_management.update_dlu_package(package_id, { "globus_dlu_status": error_msg })
+                self.data_management.update_dlu_package(package_id, { "globus_dlu_status": error_msg })
                 continue
             
             if directory_info.file_count == 0 and directory_info.subdir_count == 1:
@@ -82,9 +82,9 @@ class DLUWatcher:
               
             self.dlu_file_handler.copy_files(package_id, self.process_file_paths(directory_info.file_details))
             self.dlu_file_handler.chown_dir(package_id, file_list)
-            self.dlu_management.insert_dlu_files(package_id, file_list)
-            self.dlu_management.update_dlu_package(package_id, { "globus_dlu_status": "success" })
-            self.dlu_management.update_dlu_package(package_id, { "ready_to_move_from_globus": "done" })
+            self.data_management.insert_dlu_files(package_id, file_list)
+            self.data_management.update_dlu_package(package_id, { "globus_dlu_status": "success" })
+            self.data_management.update_dlu_package(package_id, { "ready_to_move_from_globus": "done" })
             self.dlu_mongo.update_package_files(package_id, file_list)
             self.dlu_state.set_package_state(package_id, PackageState.UPLOAD_SUCCEEDED)
             self.dlu_state.clear_cache()
